+++ tests/unit/con_peer_list.c
@@ -1,0 +1,38 @@
+/*
+ *  Turris:Sentinel Proxy - Main MQTT gateway to Sentinel infrastructure
+ *  Copyright (C) 2021 CZ.NIC z.s.p.o. (https://www.nic.cz/)
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#include <check.h>
+#include <errno.h>
+
+START_TEST(init_cleanup) {
+	ck_assert_int_eq(errno, 0);
+}
+END_TEST
+
+void unittests_add_suite(Suite*);
+
+__attribute__((constructor))
+static void suite() {
+	Suite *suite = suite_create("con_peer_list");
+
+	TCase *basic_case = tcase_create("basic");
+	tcase_add_test(basic_case, init_cleanup);
+	suite_add_tcase(suite, basic_case);
+
+	unittests_add_suite(suite);
+}