+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - logc - logging support
 - connected ZMQ peers monitoring
+- check - unit tests
+- cppcheck - linting
+- lcov - code coverage
+- valgrind
 
 ## [1.4] - 2021-05-31
 ### Added